% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trajectory_through_time}
\alias{plot_trajectory_through_time}
\title{Plot a through-time trajectory surface}
\usage{
plot_trajectory_through_time(
  x,
  metric = "ltt",
  component = "tot",
  series = NULL,
  title = NULL
)
}
\arguments{
\item{x}{A \code{trajectory_tree}.}

\item{metric}{Through-time metric.}

\item{component}{Metric component.}

\item{series}{Optional series names to retain.}

\item{title}{Optional plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a through-time trajectory surface
}
