% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{trajectory_through_time}
\alias{trajectory_through_time}
\title{Extract through-time surfaces}
\usage{
trajectory_through_time(x, metric = NULL, component = NULL)
}
\arguments{
\item{x}{A \code{trajectory_tree}, \code{trajectory_distribution}, or list of either.}

\item{metric}{Optional metric such as \code{"ltt"}, \code{"stt"}, \code{"tt"}, \code{"ctt"},
\code{"ldif"}, \code{"puniq"}, or \code{"tlen"}.}

\item{component}{Optional component such as \code{"tot"}, \code{"state"}, \code{"event"},
\code{"arr"}, \code{"leave"}, or \code{"trans"}.}
}
\value{
The requested through-time surface.
}
\description{
Extract through-time surfaces
}
