% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryTrees-package.R
\docType{package}
\name{TrajectoryTrees-package}
\alias{TrajectoryTrees}
\alias{TrajectoryTrees-package}
\title{TrajectoryTrees: evolutionary trajectory trees from stochastic maps}
\description{
The package compiles its native implementation during installation and
registers every native entry point through Rcpp.
}
\author{
\strong{Maintainer}: Sean McHugh \email{sean.mchugh4@gmail.com}

}
\keyword{internal}
